ajout liste attributions avec la possibilité de supprimer et contrôle sur l'ajout pour éviter certains soucis

+++ views/computer/listComputers.php
@@ -23,7 +23,7 @@
                                 <div>
                                     <form class="d-flex justify-content-around">
                                         <a class="btn btn-info" href="/ordinateur/modifier?id=<?php echo $value['id'] ?>">Modifier</a>
-                                        <a class="btn btn-info" href="/ordinateur/supprimer?id=<?php echo $value['id'] ?>">Supprimer</a>
+                                        <a class="btn btn-info" href="/ordinateur/supprimer?id=<?php echo $value['id'] ?>" onclick="return confirm('Voulez vous vraiment supprimer cette attribution ?')">Supprimer</a>
                                         <!-- <button class="deleteUser btn btn-info" type="submit" name="delete" value="<?php // echo $value['id'] 
                                                                                                                         ?>">Supprimer</button> -->
                                     </form>
